frontend/requirements.txt
fastapi
openai
python-multipart
python-dotenv
psycopg2-binary
duckduckgo-search
pypdf
matplotlib
python-docx
jinja2
